+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [GeoSphere AT] Fix missing info in warnings
 - [GeoSphere AT] Use the newer better endpoint for air quality
 - [JMA] Added Thai translations (@chunshek)
+- [NCEI] Added support for normals (@chunshek)
 - [Pirate Weather] Add support for daily/hourly summaries
 
 **Translations**